Implemented launcher slots.
Launcher slots function similarly to weapon or nonweapon slots, with the
exception that each launcher that requires any number of device slots will
only require ONE launcher slot, regardless of how many device slots it may
require. Ships can have the number of launcher slots set by using the
property 'maxLaunchers'; if not provided the default is one (as it is
currently). If only one launcher slot is available to a ship, then
launcher installation functions as it did previously.

The reason for this behaviour is so that behaviour of launcher
installation will function exactly as it did previously if maxLaunchers ==
1. In other words, for back compatibility.