Add new item event: GetHUDName
This event functions similar to GetName, but only affects the name as
displayed in the weapon display in the player's HUD. This means that
it has no effect on ItmGetName or similar functions.

We can use this in order to describe statuses specific to selected
weapons, without changing the gun's actual name.