In glowmaps, use 1.0 to represent distance of zero and 0 for max distance
By inverting glowmap values, if we end up having any pixels where the
alpha is nonzero due to rounding errors, the default pixel value will be
zero, which will represent maximum distance from an opaque pixel. This
avoids artifacts that we will get if we use pixel value of 0.0 to
represent distance of zero.