Add new ship AI setting: 'useAllPrimaryWeapons'.
This enables ships with multiple primary weapons installed to fire all of
them simultaneously. This is not the same as linked fire because each
weapon is fired individually by the AI.

The main use case of this is to enable ships with multiple primary weapons
to be able to aim them properly while accounting for range, which cannot
be done using secondary weapons, while not being forced to only fire one
gun at a time.